Correction in the 'extractBD' function
Namil Kim and I are working on parsing all of the 10-K forms for all the SEC registered companies (based on CIK codes), and we are using your package to retrieve the 'Business Description' section of the forms (see j1yoo4.github.io).

In the process, we have noticed that some 10-K forms use different patterns for different sections (for example, item 1. and item 2 - title) which leads the package to skip the url (for 10-K form) without retrieving the BD section (see for instance, https://www.sec.gov/Archives/edgar/data/1040721/000107997311000795/apll_2011-10k.htm).